Move CLI error helpers to shared utilities
Relocate src/cli/cli-errors.js to src/shared/cli-errors.js so CLI scripts and the prettier wrapper share one error formatting source without reaching into another workspace.
