+++ src/cli/prettier-wrapper.js
@@ -7,7 +7,11 @@ import { lstat, readdir, readFile, stat, writeFile } from "node:fs/promises";
 import { asArray } from "../shared/array-utils.js";
 import { normalizeStringList } from "../shared/string-utils.js";
 
-import { CliUsageError, formatCliError, handleCliError } from "./cli-errors.js";
+import {
+    CliUsageError,
+    formatCliError,
+    handleCliError
+} from "../shared/cli-errors.js";
 
 const wrapperDirectory = path.dirname(fileURLToPath(import.meta.url));
 const pluginPath = path.resolve(
