+++ src/shared/cli-errors.js
